+++ CHANGELOG.md
@@ -3,9 +3,19 @@ Changelog
 
 ## develop
 
+Improvements:
+
+  - [Application] Improved status (show current version) #481
+
 Bug fixes:
 
   - [Vim Plugin] Goto reference in a modified file causes warning #477.
+  - [Application] Overridden CWD not being passed to `Paths` (affected config
+    file resolution).
+
+## master
+
+  - [Docs] Added Development section
 
 ## 2018-05-20 0.5.0